results/rebuttal_new_baselines/run-20231112-235433_LLMatic-ZeroShot_donnemartin-system-design-oop-bookclub2_0_20-runs_log/bookclub2/13/LLMatic/user.py
class User:
	def __init__(self, username, password, email):
		self.username = username
		self.password = password
		self.email = email
		self.reading_list = []
		self.recommendations = []

	def create_user(self, username, password, email):
		self.username = username
		self.password = password
		self.email = email

	def update_user(self, username=None, password=None, email=None):
		if username:
			self.username = username
		if password:
			self.password = password
		if email:
			self.email = email

	def delete_user(self):
		self.username = None
		self.password = None
		self.email = None
		self.reading_list = None
		self.recommendations = None
